# fix(loader): Minimize memory size